backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
h3==4.1.2
clickhouse-connect==0.13.0
requests==2.32.3
